# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Nonprofit Operating System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Process {
  id: number;
  number: number;
  category: string;
  processDescription: string;
  processName: string;
  aiAgent: string;
  purpose: string;
  inputs: string;
  outputs: string;
  humanInTheLoop: string;
  kpi: string;
  estimatedValueImpact: string;
  industryBenchmark: string;
  included: boolean;
  target: string;
  achievement: string;
}

export interface CreateProcessBody {
  category: string;
  processDescription: string;
  processName?: string;
  aiAgent?: string;
  purpose?: string;
  inputs?: string;
  outputs?: string;
  humanInTheLoop?: string;
  kpi?: string;
  estimatedValueImpact?: string;
  industryBenchmark?: string;
  included?: boolean;
  target?: string;
  achievement?: string;
}

export interface UpdateProcessBody {
  category?: string;
  processDescription?: string;
  processName?: string;
  aiAgent?: string;
  purpose?: string;
  inputs?: string;
  outputs?: string;
  humanInTheLoop?: string;
  kpi?: string;
  estimatedValueImpact?: string;
  industryBenchmark?: string;
  included?: boolean;
  target?: string;
  achievement?: string;
}

export interface DeleteResponse {
  success: boolean;
}

export interface ImportResult {
  inserted: number;
  updated: number;
  total: number;
}

export interface ErrorResponse {
  error: string;
}

export interface AuditLog {
  id: number;
  timestamp: string;
  action: string;
  entityType: string;
  entityId?: string | null;
  entityName?: string | null;
  fieldChanged?: string | null;
  oldValue?: string | null;
  newValue?: string | null;
  user: string;
  description?: string | null;
}

export interface CreateAuditLogBody {
  action: string;
  entityType: string;
  entityId?: string;
  entityName?: string;
  fieldChanged?: string;
  oldValue?: string;
  newValue?: string;
  user?: string;
  description?: string;
}

export interface AnthropicConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface AnthropicMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateAnthropicConversationBody {
  title: string;
}

export interface SendAnthropicMessageBody {
  content: string;
}

export interface AnthropicConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: AnthropicMessage[];
}

export interface AnthropicError {
  error: string;
}

export type ImportProcessesBody = {
  file?: Blob;
};

export type ListAuditLogsParams = {
  limit?: number;
};