# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nonprofit Operating System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: processes
    description: Nonprofit process operations
  - name: anthropic
    description: Anthropic AI chat operations
  - name: audit
    description: Audit log operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /processes:
    get:
      operationId: listProcesses
      tags: [processes]
      summary: List all processes
      responses:
        "200":
          description: List of processes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Process"
    post:
      operationId: createProcess
      tags: [processes]
      summary: Create a new process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessBody"
      responses:
        "201":
          description: Created process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /processes/export:
    get:
      operationId: exportProcesses
      tags: [processes]
      summary: Export processes to Excel
      responses:
        "200":
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: {}

  /processes/{id}:
    get:
      operationId: getProcess
      tags: [processes]
      summary: Get a process by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Process found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateProcess
      tags: [processes]
      summary: Update a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProcessBody"
      responses:
        "200":
          description: Updated process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProcess
      tags: [processes]
      summary: Delete a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /processes/{id}/ai-populate:
    post:
      operationId: aiPopulateProcess
      tags: [processes]
      summary: Use AI to populate blank fields for a process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: AI-populated process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Process"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      operationId: listCategories
      tags: [processes]
      summary: List distinct categories
      responses:
        "200":
          description: List of category names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /processes/import:
    post:
      operationId: importProcesses
      tags: [processes]
      summary: Import processes from Excel (upsert by number)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit]
      summary: List audit log entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: List of audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
    post:
      operationId: createAuditLog
      tags: [audit]
      summary: Create an audit log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAuditLogBody"
      responses:
        "201":
          description: Created audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLog"

  /anthropic/conversations:
    get:
      operationId: listAnthropicConversations
      tags: [anthropic]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicConversation"
    post:
      operationId: createAnthropicConversation
      tags: [anthropic]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnthropicConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversation"

  /anthropic/conversations/{id}:
    get:
      operationId: getAnthropicConversation
      tags: [anthropic]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"
    delete:
      operationId: deleteAnthropicConversation
      tags: [anthropic]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicError"

  /anthropic/conversations/{id}/messages:
    get:
      operationId: listAnthropicMessages
      tags: [anthropic]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnthropicMessage"
    post:
      operationId: sendAnthropicMessage
      tags: [anthropic]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendAnthropicMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Process:
      type: object
      properties:
        id:
          type: integer
        number:
          type: integer
        category:
          type: string
        processDescription:
          type: string
        processName:
          type: string
        aiAgent:
          type: string
        purpose:
          type: string
        inputs:
          type: string
        outputs:
          type: string
        humanInTheLoop:
          type: string
        kpi:
          type: string
        estimatedValueImpact:
          type: string
        industryBenchmark:
          type: string
        included:
          type: boolean
        target:
          type: string
        achievement:
          type: string
      required:
        - id
        - number
        - category
        - processDescription
        - processName
        - aiAgent
        - purpose
        - inputs
        - outputs
        - humanInTheLoop
        - kpi
        - estimatedValueImpact
        - industryBenchmark
        - included
        - target
        - achievement

    CreateProcessBody:
      type: object
      properties:
        category:
          type: string
        processDescription:
          type: string
        processName:
          type: string
        aiAgent:
          type: string
        purpose:
          type: string
        inputs:
          type: string
        outputs:
          type: string
        humanInTheLoop:
          type: string
        kpi:
          type: string
        estimatedValueImpact:
          type: string
        industryBenchmark:
          type: string
        included:
          type: boolean
        target:
          type: string
        achievement:
          type: string
      required:
        - category
        - processDescription

    UpdateProcessBody:
      type: object
      properties:
        category:
          type: string
        processDescription:
          type: string
        processName:
          type: string
        aiAgent:
          type: string
        purpose:
          type: string
        inputs:
          type: string
        outputs:
          type: string
        humanInTheLoop:
          type: string
        kpi:
          type: string
        estimatedValueImpact:
          type: string
        industryBenchmark:
          type: string
        included:
          type: boolean
        target:
          type: string
        achievement:
          type: string

    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    ImportResult:
      type: object
      properties:
        inserted:
          type: integer
        updated:
          type: integer
        total:
          type: integer
      required:
        - inserted
        - updated
        - total

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AuditLog:
      type: object
      properties:
        id:
          type: integer
        timestamp:
          type: string
          format: date-time
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
          nullable: true
        entityName:
          type: string
          nullable: true
        fieldChanged:
          type: string
          nullable: true
        oldValue:
          type: string
          nullable: true
        newValue:
          type: string
          nullable: true
        user:
          type: string
        description:
          type: string
          nullable: true
      required:
        - id
        - timestamp
        - action
        - entityType
        - user

    CreateAuditLogBody:
      type: object
      properties:
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: string
        entityName:
          type: string
        fieldChanged:
          type: string
        oldValue:
          type: string
        newValue:
          type: string
        user:
          type: string
        description:
          type: string
      required:
        - action
        - entityType

    AnthropicConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    AnthropicMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    CreateAnthropicConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendAnthropicMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    AnthropicConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    AnthropicError:
      type: object
      properties:
        error:
          type: string
      required:
        - error